rag_pipeline/pyproject.toml
[project]
name = "rag-pipeline"
version = "0.1.0"
description = "RAG Pipeline with OpenAI Agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.6.4",
    "openai>=1.0.0",
    "cohere>=5.20.0",
    "qdrant-client>=1.16.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.14.0",
]